Refactor: modify controller name, separate login logic to loginController
