+++ routes/controller/condition.js
@@ -11,7 +11,6 @@ const Grid = require("../../models/CustomGrid");
 const getISOTime = require("../utils/getISOTime");
 const ACCESS_LEVELS = require("../../constants/accessLevels");
 const { OK, UNAUTHORIZED } = require("../../constants/statusCodes");
-const { generateToken, verifyToken } = require("../helpers/tokens");
 
 async function getCondition(req, res, next) {
   try {
